+++ src/lib/api/rule34xxx/index.ts
