+++ src/lib/api/rule34xxx/rule34xxx.interface.ts
